+++ lib/tree_view.dart
@@ -1,8 +1,7 @@
 library flutter_treeview;
 
+export 'src/expander_theme_data.dart';
 export 'src/models/node.dart';
-export 'src/models/node_icon.dart';
 export 'src/tree_view.dart';
 export 'src/tree_view_controller.dart';
 export 'src/tree_view_theme.dart';
-export 'src/expander_theme_data.dart';